// Application Code/frontend/src/app/models/signup.ts
export interface StudentSignup {
  firstname: string;
  lastname: string;
  uniqueid: number;
  password: string;
  course: number;
  department: number;
  semester: number;
  profession: number;
}
export interface TeacherSignup {
  firstname: string;
  lastname: string;
  uniqueid: number;
  password: string;
  department: number;
  profession: number;
}